+++ data/miscanthus_x_giganteus.R
@@ -12,7 +12,8 @@ miscanthus_x_giganteus <- list(
         "BioCro:stefan_boltzmann_longwave",
         canopy_photosynthesis = "BioCro:c4_canopy",
         partitioning_coefficients = "BioCro:partitioning_coefficient_selector",
-        partitioning_growth_calculator = "BioCro:partitioning_growth_calculator"
+        partitioning_growth_calculator = "BioCro:partitioning_growth_calculator",
+        "BioCro:carbon_assimilation_to_biomass"
     ),
     differential_modules = list(
         senescence = "BioCro:thermal_time_senescence",
@@ -58,11 +59,13 @@ miscanthus_x_giganteus <- list(
         beta                        = 0.93,
         Catm                        = 400,
         chil                        = 1,
+        dry_biomass_per_carbon      = 30.026, # g CHO / mol C (glucose)
         emissivity_sky              = 1,
         Gs_min                      = 1e-3,
         heightf                     = 1.33, # LAI of 4 when canopy is 3 m tall
         hydrDist                    = 0,
         iSp                         = 1.7,
+        k_diffuse                   = 0.1,
         kGrain1                     = 0,
         kGrain2                     = 0,
         kGrain3                     = 0,
@@ -96,15 +99,14 @@ miscanthus_x_giganteus <- list(
         kStem4                      = 0.63,
         kStem5                      = 0.63,
         kStem6                      = 0.63,
-        k_diffuse                   = 0.1,
         lat                         = 40,
-        LeafN                       = 2,
-        LeafN_0                     = 2,
-        leafwidth                   = 0.04,
         leaf_reflectance_nir        = 0.38, # Corn values from Table 7-1 from Norman & Arkebauer (1991) https://doi.org/10.2134/agronmonogr31.c7
         leaf_reflectance_par        = 0.09, # Corn values from Table 7-1 from Norman & Arkebauer (1991) https://doi.org/10.2134/agronmonogr31.c7
         leaf_transmittance_nir      = 0.45, # Corn values from Table 7-1 from Norman & Arkebauer (1991) https://doi.org/10.2134/agronmonogr31.c7
         leaf_transmittance_par      = 0.04, # Corn values from Table 7-1 from Norman & Arkebauer (1991) https://doi.org/10.2134/agronmonogr31.c7
+        LeafN                       = 2,
+        LeafN_0                     = 2,
+        leafwidth                   = 0.04,
         lnfun                       = 0,
         longitude                   = -88,
         lowerT                      = 3,
@@ -140,8 +142,8 @@ miscanthus_x_giganteus <- list(
         seneRoot                    = 4000,
         seneStem                    = 3500,
         soil_air_entry              = -2.6,
-        soil_bulk_density           = 1.35,
         soil_b_coefficient          = 5.2,
+        soil_bulk_density           = 1.35,
         soil_clay_content           = 0.34,
         soil_clod_size              = 0.04,
         soil_depth1                 = 0.0,
@@ -156,8 +158,8 @@ miscanthus_x_giganteus <- list(
         soil_transmission           = 0.01,
         soil_wilting_point          = 0.2,
         sowing_fractional_doy       = 0,
-        specific_heat_of_air        = 1010,
         Sp_thermal_time_decay       = 0,
+        specific_heat_of_air        = 1010,
         tbase                       = 0,
         theta                       = 0.83,
         timestep                    = 1,