+++ NEWS.md
@@ -127,6 +127,15 @@ be directly added to this file to describe the related changes.
   `update_stored_model_results`. Previously, these were part of
   `tests/testthat/crop_model_testing_helper_functions.R`.
 
+- The conversion of CO2 assimilation to biomass is no longer hard coded into
+  the photosynthesis modules: c3CanAC, CanAC, multilayer_canopy_integrator, etc.
+  These modules now produce canopy assimilation rates in micromol CO2 / m^2 / s.
+  A new module called `BioCro:carbon_assimilation_to_biomass` now performs the
+  conversion. A new parameter `dry_biomass_per_carbon` controls the conversion.
+  All affected models have the same behavior as before if the new module is used
+  with `dry_biomass_per_carbon = 30.026` g / mol. See module documentation for
+  details.
+
 ## OTHER CHANGES
 
 - Consolidated all temperature response functions into a single header file